+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Added the `php craft db/modify-table-prefix` command, which can be used to change the table prefix for an existing Craft install.
 - Addresses’ owner elements are now automatically set on them during initialization, if they were queried with the `owner` address query param.
 - Entry Title fields are no longer shown when “Show the Title field” is disabled and there’s a validation error on the `title` attribute. ([#13876](https://github.com/craftcms/cms/issues/13876))